Add Screenshot of Model
This is the first model built.
I've added all variables to the model to determine which are to be eliminated from the model to make it more accurate.

I have established SL at 5% or 0.05. Any variable with the highest p-value above 0.05 will be eliminated and the model re-run through Gretl without the variable with high p-value.

In this model, the variable Administration has the highest p-value at 0.65 well above the threshold of 0.05, thus identifying it statistically having a weak correlation to the dependent variable (in this case profit).