Add Third Model
This model labels the variable Marketing Expenses with the highest p-value. However the p-value is close enough to our 5 % that we decide to give it a second look.

When marketing against profit is graphed there does seem to be a correlation because as marketing dollars go up, so do profits.
We may decide to keep Marketing in the model.